requirements.txt
Flask==1.1.1
numpy==1.17.4
pandas==0.25.3
Pillow==6.2.1
requests==2.22.0
tqdm==4.43.0

--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.4.0+cpu 
torchvision==0.5.0+cpu